CEA_mongo/requirements.txt
pymongo==4.3.3
werkzeug==2.3.4
Flask-Bcrypt==1.0.1